requirements.txt
accelerate==1.1.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.1.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.5
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
av==13.1.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==1.0.1
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
clip==1.0
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.8
datasets==3.1.0
debugpy==1.6.7
decorator==5.1.1
deepseek_vl==1.0.0
deepseek_vl2==1.0.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
editdistance==0.8.1
einops==0.8.0
en_core_web_sm==3.8.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
faiss-cpu==1.9.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flash-attn==2.7.0.post2
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gdown==5.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==2.7.3
googleapis-common-protos==1.68.0
groundingdino-py==0.4.0
grpcio==1.70.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.26.2
hyperframe==6.1.0
icecream==2.1.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.5.1
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistral_common==1.5.3
mistral_inference==1.5.0
mistune==3.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open_clip_torch==2.29.0
openai==1.0.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
peft==0.13.2
pexpect==4.8.0
pillow==11.0.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
preshed==3.0.9
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.2.0
proto-plus==1.26.0
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.15.1
pyparsing==3.2.0
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==24.0.1
qwen-vl-utils==0.0.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment_anything==1.0
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-huggingface-pipelines==0.0.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacyface==0.1.5
srsly==2.4.8
stack-data==0.2.0
supervision==0.6.0
sympy==1.13.1
termcolor==2.5.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typer==0.13.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.11.0
typing_utils==0.1.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
word2number==1.1
wrapt==1.16.0
xformers==0.0.29.post1
xxhash==3.5.0
yapf==0.43.0
yarl==1.17.1
zipp==3.21.0
zstandard==0.23.0
